[fix] 좋아요 branch 분리
