+++ src/routes/index.js
@@ -1,12 +1,12 @@
 const express = require('express');
 const router = express.Router();
 
-const userRouter = require("./userRouter");
+const confirmationsRouter = require("./confirmationsRouter");
 
 router.get('/', (req, res, next) => {
   res.send('OK');
 });
 
-router.use('/mypage', userRouter);
+router.use('/confirmations', confirmationsRouter);
 
 module.exports = router;
