+++ src/services/confirmationsService.js
@@ -1,0 +1,9 @@
+const { userDao } = require('../models');
+
+const getLikedEvents = async (userId) => {
+  return await userDao.userLike(userId);
+};
+
+module.exports = {
+  getLikedEvents,
+};
